+++ fiddles/angular2/template/typings/tsd.d.ts
@@ -1,3 +1,0 @@
-/// <reference path="es6-promise/es6-promise.d.ts" />
-/// <reference path="rx/rx-lite.d.ts" />
-/// <reference path="rx/rx.d.ts" />