+++ fiddles/bash/fiddle-0064-InstallTypings/test.sh
@@ -1,0 +1,25 @@
+#!/usr/bin/env bash
+
+source script.sh
+
+function catch() {
+    case $1 in
+        0)  echo "all tests succeeded"
+            ;;
+        1)  echo "installTypings() failed";
+            ;;
+        *)  echo "fubar! Something went wrong."
+            ;;
+    esac
+    exit $1
+}
+# try
+(
+    installTypings || exit 1
+)
+catch $?;
+
+
+
+
+